+++ app/src/main/java/com/example/justwork/viewmodel/AccountViewModel.java
@@ -19,7 +19,7 @@ public class AccountViewModel extends AndroidViewModel {
         super(application);
         accountRepository = AccountRepository.getInstance(application);
     }
-    public void registerUser(int cpr, String username, String email, String password, int phoneNumber,
+    public void registerUser(long cpr, String username, String email, String password, int phoneNumber,
                              String address, DrivingLicenceList drivingLicences, String gender, String nationality){
         accountRepository.registerUser(cpr, username, email, password, phoneNumber, address, drivingLicences, gender, nationality);
     }
